+++ source/extensions/omni.isaac.debug_draw/nodes/OgnIsaacXPrimAxisVisualizer.cpp
@@ -12,6 +12,7 @@
 // clang-format off
 #include "UsdPCH.h"
 // clang-format on
+#include <omni/fabric/FabricUSD.h>
 #include <omni/isaac/debug_draw/PrimitiveDrawingHelper.h>
 #include <omni/isaac/utils/BaseResetNode.h>
 #include <omni/isaac/utils/Pose.h>
@@ -50,7 +51,18 @@ public:
 
     static bool compute(OgnIsaacXPrimAxisVisualizerDatabase& db)
     {
-        auto primPath = pxr::SdfPath(db.inputs.xPrim.path());
+        const auto& prim = db.inputs.xPrim();
+        pxr::SdfPath primPath;
+        if (prim.size() > 0)
+        {
+            primPath = omni::fabric::toSdfPath(prim[0]);
+        }
+        else
+        {
+            db.logError("no prim path found");
+            return false;
+        }
+
         auto& state = db.internalState<OgnIsaacXPrimAxisVisualizer>();
         state.mLength = db.inputs.length();
         state.mThickness = db.inputs.thickness();